Remove GetDefaultInitialCandidateObjects (also removed doubt about SetTopLevelContent)
== GetDefaultInitialCandidateObjects ==

`GetDefaultInitialCandidateObjects` narrows the search space of objects when no list of candidates is provided.

If a single object or a containers of matches and non-matches are passed to `Eval`, then just those objects are tested by the condition. But there is an `Eval` overload that just takes a set of matches, which it immediately clears, and then fills with any objects in the universe that the condition matches.

To make this faster, conditions have the `GetDefaultInitialCandidateObjects` function to initialize the search space of objects to test, rather than always testing all objects in the `Universe`. For example, `Source` initializes with just the `ScriptingContext`'s source object, and `Type` will initialize based on what object type is being matched (eg. ships for `OBJ_SHIP`).

For this case, it's acting like an `Or` condition, and thus shouldn't narrow the search space at all, so like `Or` you shouldn't overload `GetDefaultInitialCandidateObjects` and instead just use the base class behaviour.

== SetTopLevelContent ==

`SetTopLevelContent` is used to set `m_top_level_content`, which is declared here:

https://github.com/freeorion/freeorion/blob/master/universe/ValueRef.h#L207

and which is used here:

https://github.com/freeorion/freeorion/blob/master/universe/ValueRef.cpp#L552

to return the top level bit of content in which this ValueRef is scripted. For example, in a species script, it would be the name of the species, returned as a `ValueRef::Constant<std::string>` when `ThisSpecies` is written into the script. This is helpful for writing generic macros that can be used in different contexts.

All the Effects, Conditions, and ValueRefs all have `SetTopLevelContent` functions, as they can contain eachother recursively and the final leaf ValueRef needs to have its top level content set by its enclosing object.