Add ShipDesignsInProduction ValueRef and use it for upkeep
ShipDesignsInProduction returns the (count of) empire's ship batches in the production queue
if work on that project already started i.e. if you already spent PP on it.

NOTE: As we do not (yet) count the ship parts in the queue, part upkeep assumes the same cost
per ship in the queue as for ship based upkeep (equating to four ship parts).
This makes upkeep especially for unfinished decoys quite more expensive than for finished decoys.
It might make sense to penalize keeping unfinished projects in the queue.

This also calls UpdateOwnedObjectCounters when loading the game.
It is possible that other update calls are missing e.g. for statistics