Add ShipDesignsInProduction ValueRef and use it for upkeep
ShipDesignsInProduction returns the (count of) empire's ship batches in the production queue
if work on that project already started i.e. if you already spent PP on it.

NOTE: As we do not (yet) count the ship parts in the queue, part upkeep assumes one ship part
per ship in the queue. This makes upkeep for unfinished decoys actually more expensive than
for finished decoys. It might make sense to penalize keeping unfinished projects in the queue.

This also calls UpdateOwnedObjectCounters when loading the game.
It is possible thar other update calls are missing e.g. for statistics