Let METER_MAX_FUEL hull effect execute very very late (prio 500)
This makes it possible to remove the workaround in ship_hulls.macros
and the effect accounting order is more sensible.

For the perfect UI BAD_FUEL species effect should also be shifted to positive values.

Alternatives - change in UI:
Use own FuelMeterBrowseWnd in ShipDetailPanel (FleetWnd)
or modify GetAccountingInfo in MeterBrowseWnd
            - change order of accountingInfos when those are created (Universe)
i had some crashes and mistakenly thought the inherent would be the meter based entry