NamedValueRefParser working for SHP_REINFORCED_HULL_BONUS (double value ref)
* NamedValueRefParser can parse one type of value ref (double, which is the most important)
** would also work for integer, if the parser would not get confused in the second parse branch(?)
** works in pedia (dont forget you need to add a stringtable entry for the value tag to work)
** works in the game
* moved GetValueRef to header so its visible

Still Named-in-the-middle should be made to work..
e.g. the SHP_REINFORCED_HULL_BONUS value is a good example for a value should be defined in the tech file
It does not make sense to have all that definitions in named_values (and it will be a lot) instead of the place
where you would actually need/use those.