Fix estimated monster fighter damage by introducing DUMMY_EVIL_EMPIRE id
The shown ship damage estimates for fleets, ships and fighter per bout take the targeting condition into account.
This did not work for monsters as the target candidate was not an enemy of neutral.

This adds DUMMY_EVIL_EMPIRE -2 to dummy ships used for determining valid targets.
For the targeting condition, the empire affiliations are checked.
The DUMMY_EVIL_EMPIRE is at war with every real empire and also neutral forces.
It is not at peace/alliance with anyone. It is not a human player empire.

While currently irrelevant, it is not completely clear
what the right values for AFFIL_ANY and AFFIL_NONE should be.
Currently for those, it is treated like a normal empire.