* count Colony parts quadruple complexity
Implementation:
* DESIGN_SIMPLICITY_SOURCE_COMPLEXITY_COUNT_VREF checks the ship design for each part type and counts on hit (colony parts -> +4)
* DESIGN_SIMPLICITY_COMPLEXITY_FACTOR_FOR_ARG1_VREF has a if-else cascade each rechecking the previous count and returning the discount
* stringtables
* AI (no changes necessary for basic support)

Missing:
* performance testing

Notes
* both macros seem quite inefficient and have to be calculated for every ship design when opening the build queue. hoping caching/relying on invariants is good enough
* XXX maybe use a NamedValue for each design or store the complexity in the ship design itself in order to prevent recalculation
* part names are no objects; so this does not work (3 > Statistic UniqueCount value = LocalCandidate. candidate = CONDITION_RESULT_IS_OBJECTS
* ...cant discriminate part names ; there is no support for list type attributes

== AI ==
If it worked before, it should still work.

PolicyAI has a heuristic how good design simplicity is - that looks fine.

ShipDesigner should estimate ship design based on the real values with the applied policy,
so the simpler designs should be comparatively cheaper and the AI should use designs
based on efficiency afterwards.

The named values could be used in the AI, but are currently unused.